requirements.txt
google-generativeai
google-genai
streamlit
python-dotenv